+++ crates/objc2/src/runtime/mod.rs
@@ -32,7 +32,7 @@ use std::os::raw::c_uint;
 #[doc(hidden)]
 pub mod __nsstring;
 mod bool;
-pub(crate) mod message;
+mod message_receiver;
 mod method_encoding_iter;
 mod method_implementation;
 mod nsobject;
@@ -53,7 +53,7 @@ use crate::{ffi, Message};
 pub use self::nsproxy::NSProxy as __NSProxy;
 
 pub use self::bool::Bool;
-pub use self::message::{MessageArguments, MessageReceiver};
+pub use self::message_receiver::MessageReceiver;
 pub use self::method_implementation::MethodImplementation;
 pub use self::nsobject::{NSObject, NSObjectProtocol};
 pub use self::nszone::NSZone;